clean up

+++ windows/cross-compile/build.sh
@@ -883,24 +883,27 @@ fi
 # Toxcore
 
 TOXCORE_PREFIX_DIR="$DEP_DIR/libtoxcore"
-TOXCORE_HASH=not_used
 TOXCORE_CUSTOM_VERSION=${TOXCORE_CUSTOM_VERSION:-}
 
 if [ "$TOXCORE_CUSTOM_VERSION""x" == "x" ]; then
   TOXCORE_REPO="https://github.com/TokTok/c-toxcore"
   TOXCORE_VERSION="v0.1.11"
+  TOXCORE_HASH="518b7b3e8710943561b8c916b8e11f36096e1820add9dad3a7cc61c70240409f"
 else
-  echo "Using custom c-toxcore repo:$TOXCORE_REPO version:$TOXCORE_VERSION"
+  echo "Using custom c-toxcore repo:$TOXCORE_REPO version:$TOXCORE_VERSION hash:$TOXCORE_HASH"
 fi
 
-#if [ ! -f "$TOXCORE_PREFIX_DIR/done" ]
-#then
+if [ ! -f "$TOXCORE_PREFIX_DIR/done" ]
+then
   rm -rf "$TOXCORE_PREFIX_DIR"
   mkdir -p "$TOXCORE_PREFIX_DIR"
   git clone "$TOXCORE_REPO"
   cd c-toxcore
   git checkout "$TOXCORE_VERSION"
-
+  # do not check hash for a custom version of c-toxcore
+  if [ "$TOXCORE_CUSTOM_VERSION""x" == "x" ]; then
+    check_sha256_git "$TOXCORE_HASH"
+  fi
   mkdir -p build
   cd build
 
@@ -937,9 +940,9 @@ fi
 
   cd ..
   rm -rf ./c-toxcore
-#else
-#  echo "Using cached build of Toxcore `cat $TOXCORE_PREFIX_DIR/done`"
-#fi
+else
+  echo "Using cached build of Toxcore `cat $TOXCORE_PREFIX_DIR/done`"
+fi
 
 
 # mingw-w64-debug-scripts
